% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing.R
\name{add_missing}
\alias{add_missing}
\title{Adding Missing Entries to a Matrix}
\usage{
add_missing(data, mask)
}
\arguments{
\item{data}{Matrix with numeric entries that is to be masked.}

\item{mask}{Binary matrix indicating where missing entries should be imposed (0 = missing, 1 = not missing)}
}
\value{
A numeric matrix with missing entries given by \code{NA}
}
\description{
Adding Missing Entries to a Matrix
}
