% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amount_missing.R
\name{amount_missing}
\alias{amount_missing}
\title{Extracting the Amount of Missing Entries}
\usage{
amount_missing(data)
}
\arguments{
\item{data}{Matrix with numeric entries and missing entries given as \code{NA}.}
}
\value{
A vector with the column-wise fraction of missing entries.
}
\description{
Internal function of ImputeBench.
}
