% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_missingness_plots.R
\name{analyse_missingness_plots}
\alias{analyse_missingness_plots}
\title{ImputeBench Internal Plot Function}
\usage{
analyse_missingness_plots(
  data,
  imputed.data = NULL,
  clms = NULL,
  mask = NULL,
  error.groups = NULL,
  missing.label = "percent",
  first.clms.entry.missing = FALSE,
  only.first.mask.missingness = TRUE,
  jitter = 0.4,
  alpha_set = 0.25,
  alpha_line = 1,
  seed = 1,
  title = NULL,
  thres_name = 0.05,
  binwidth = NULL,
  multiple.imputations = NULL,
  imp.names = NULL
)
}
\arguments{
\item{data}{A numeric matrix (possibly with missing entries given by \code{NA}).}

\item{imputed.data}{A numeric matrix (the imputed counterpart to \code{data}). Default is \code{NULL}.}

\item{clms}{The columns/variables that should be plotted. Default is \code{NULL}.}

\item{mask}{A binary matrix indicating additional missing entries in \code{data} (0 = missing, 1 = present). Default is \code{NULL}.}

\item{error.groups}{A list of column indices yielding a grouping of the columns. Default is NULL.}

\item{missing.label}{A character sting. How to report the amount missing, as \code{"percent"} or \code{"amount"}?}

\item{first.clms.entry.missing}{Boolean, should the data distribution be displayed or the missingness of the first entry of \code{clms} with respect
to the data distributions of the remaining entries in \code{clms}? Default is \code{FALSE}.}

\item{only.first.mask.missingness}{Boolean, should only the missingness of the first entry in clms be considered or the missingness in all entries of \code{clms}?
Default is \code{TRUE}.}

\item{jitter}{Numeric controlling the jitter added to count variables. Default is \code{0.4}.}

\item{alpha_set}{Numeric controlling the alpha/transparency of the data points (or the existing data points). Default is \code{0.25}.}

\item{alpha_line}{Numeric controlling the alpha/transparency of the lines connecting missing and imputed data points. Default is \code{1}.}

\item{seed}{Integer controlling the seed in the jitter of the labels when plotting amount of missing entries per column. Default is \code{1}.}

\item{title}{A character string: User specified title of the resulting plot. Default is \code{NULL}.}

\item{thres_name}{Numeric controlling the threshold over which labels of the column names should be plotted when plotting amount of
of missing entires per column. Default is \code{0.05}.}

\item{binwidth}{Integer controlling the bin width for histogram plots. Default is \code{NULL}, resulting in \code{ncol(data)/30}.}

\item{multiple.imputations}{List of numeric matrices: Do we compare in the 1D plots multiple imputation methods on ground truth data?
Then \code{multiple.imputations} lists the additional imputed matrices (additional to \code{imputed.data}). Default is \code{NULL}.}

\item{imp.names}{A vector of character strings. In the case a list of imputation matrices is provided in \code{multiple.imputations} this
vector can provide names for the imputation methods used, i.e. controls the color legend. Default is \code{NULL}.}
}
\value{
A plot or a list of plots.
}
\description{
ImputeBench Internal Plot Function
}
\details{
Internal function for plotting data, missing data, imputed data and missingness patterns.
}
