% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ImputeBench.R
\name{data_ImputeBench}
\alias{data_ImputeBench}
\title{Benchmarking Imputation Methods on Real Data}
\usage{
data_ImputeBench(
  data,
  methods = NULL,
  training = NULL,
  scenarios = NULL,
  missingness_parameters = NULL,
  repetitions = 5,
  error_grouping = NULL,
  default_methods = c("knn", "MICE", "missF", "softImpute"),
  baseline_method = "median",
  error_measure = "l2",
  scaling_robust = 0.01,
  retrain_always = FALSE,
  verbose = FALSE,
  seed = NULL,
  max_time = 2592000,
  output_data = TRUE
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{methods}{A (named) list of functions, taking as an input a numeric matrix with missing entries via an argument \code{data} and a list of
parameters via an argument \code{args}. Default is \code{NULL}, i.e. no additional methods added to the benchmarking.}

\item{training}{A list of functions - which need to be of the same length as the \code{methods} list -, that take as an input
a data matrix with missing entries via an argument \code{data} and outputs the parameters for the respective method via as a list \code{args}.
Default is \code{NULL}.}

\item{scenarios}{A (named) list of named lists of missingness scenarios (parameters thereof), for details and an example see below and the vignette.
If \code{NULL} scenarios are drawn by \code{missingness_parameters}. In the case also \code{missingness_parameters} is \code{NULL} a default missingness
scenario is used, see the vignette for details. Default is \code{NULL}.}

\item{missingness_parameters}{A (named) list of named lists of parameters defining the way missingness scenarios are drawn, for details see below
and the vignette. In the case \code{scenarios} are provided, \code{missingness_parameters} is ignored. Default is \code{NULL}.}

\item{repetitions}{Integer determining how often an additional missingness pattern according to one entry of \code{scenarios} or  \code{missingness_parameters}
is drawn. Default is \code{5}.}

\item{error_grouping}{A list of indice vectors defining column groups. Errors for each groups are collected and the overall error
is computed by the mean error over all groups (weighting all groups the same irrespective of the number of columns they contain).
Default is \code{NULL}.}

\item{default_methods}{\code{NULL} or vector with four possible character string entries: \code{"knn"}, \code{"MICE"}, \code{"missF"} and \code{"softImpute"}. Determines
which default imputation method is included in the comparison. Default is \code{c("knn","MICE","missF","softImpute")}.}

\item{baseline_method}{Character string, either \code{"median"} (default) or \code{"mean"}. Defining the baseline imputation method.}

\item{error_measure}{Character string that must be \code{"l2"} (up to now only error measure supported). Default is \code{"l2"}.}

\item{scaling_robust}{Positive numeric, robustness parameter avoiding division by close to zero when the columns are scaled by the standard deviation
when computing the imputation error. Default is \code{0.01}.}

\item{retrain_always}{Boolean, should parameters for each imputation method be learned for every run or only every scenario/parameter choice?
Default is \code{FALSE}.}

\item{verbose}{Boolean, determines if an output to the console is made each time a run is finished. Default is \code{FALSE}.}

\item{seed}{Integer or \code{NULL}, sets a seed. Default is \code{NULL}.}

\item{max_time}{The maximal amount of seconds an imputation method is allowed to take. Default is \code{2592000}, i.e. 30 days.}

\item{output_data}{Boolean, should additional data be collected, i.e. drawn missingness pattern? Default is \code{TRUE}.}
}
\value{
A list is returned by \code{data_ImputeBench()} with one or two entries. The first is \verb{$evaluation}
under which one finds a data frame listing imputation performance per run and method and which can be plugged into
\code{plot_ImputeBench()} for a plot of performance. The second entry is included if \code{output_data = TRUE} and is called \verb{$data.list}. One can find here
all drawn missingness pattern and all trained parameter of the imputation methods, for details see the vignette.
}
\description{
Simulation Study to compare imputation methods on a given data with additionally simulated missingness
according to the benchmarking protocol of RMTBM22.
}
\details{
This function benchmarks imputation methods on user-provided data. Via the argument \code{data} a numeric matrix must be passed
that may or may not contain missing values given as \code{NA}-entries. A detailed description of the functionality and arguments of
\code{data_ImputeBench()} can be found in the accompanying vignette. For the exact definition of the benchmarking protocol, the
missingness pattern simulation protocol and the definition of a missingness scenario we refer to the paper RMTBM22, see below.
In the following a brief overview over the arguments and example inputs.
\subsection{Customizing the Compared Imputation Methods}{

To add to the default imputation methods user specified ones \code{data_ImputeBench()} provides the argument \code{methods} which expects a
list of functions to be passed. These functions have to have exactly two inputs: \code{data} (a numeric matrix with missing entries
denoted by \code{NA}) and \code{args} (a list with parameters for the imputation method, this can be left \code{NULL}). The output needs to be a
numeric matrix without missing entries (by default still missing entries are filled by the baseline imputation method). For details
see the accompanying vignette. Customizing names of the additional methods is done by naming list entries.

Example of \code{methods}:\preformatted{methods = list("Random Imputation" =  function(data,
                         args = NULL)\{
    clms.with.missing.entries =
       which(matrixStats::colSums2(matrix(as.numeric(is.na(data)), nrow = nrow(data))) != 0)
    imputed.matrix = data
    for(t in clms.with.missing.entries)\{
       imputed.matrix[is.na(data[,t]),t] =
       sample(data[!is.na(data[,t]),t], replace = TRUE, size = sum(as.numeric(is.na(data[,t]))))
    \}
    return(imputed.matrix)
  \},
                "User Candidate 2" = function(data, args = ...)\{ ... return(imputed.matrix)\})
}

Via the argument \code{training} parameter training functions can be passed to \code{data_ImputeBench()} (note, that this is necessary
if methods are added via \code{methods} which do not except \code{args = NULL} as an input). The argument \code{training} expects as an
input similar to \code{methods} a list of functions, which take as an input only one argument \code{data} and which output the argumnet
\code{args}, a list of parameter choices. The output of functions included in \code{training} must be compatible with the input \code{args}
of the index-wise corresponding functions in \code{methods}.

Example of \code{training}:\preformatted{methods = list(  function(data)\{
    args = NULL
    return(args)
   \},
                 function(data)\{ ... return(args)\})
}

By default the parameter \code{default_methods} is set to \code{c("knn", "MICE", "missF", "softImpute")} including all four supported
default imputation methods in the benchmarking. To remove the default parameters remove them from the vector and supply the
resulting vector to \code{default_methods} (\code{NULL} in case all default methods should be removed).

By default parameter of the imputation methods are trained (default training is supported for k-nn imputation (\code{k}) and soft
impute (\code{lambda})) once every scenario by the default argument \code{retrain_always = FALSE}. To have parameters being retrained
for every run set \code{retrain_always = TRUE}.

The baseline imputation method is median imputation set via the default \code{baseline_method = "median"}. To switch to the second
supported baseline imputation method pass \code{baseline_method = "mean"} for mean imputation.
}

\subsection{Customizing the Missingness Pattern Simulation Protocol}{

The parameter \code{scenarios} expects a named list of named lists as an entry. Each entry represents a scenario under which imputation
methods are compared. Each such list entry contains at least one and up to three list entries named \code{"MCAR"}, \code{"MAR"} and/or \code{"MNAR"},
each of which contain the parameters of the missingness pattern protocol for columns with MCAR, MAR and MNAR missingness, respectively.
For the definition of a missingness scenario and the protocol under which the missingness pattern is being drawn we refer to
RMTBM22 and the accompanying vignette.

Example for \code{scenarios}:\preformatted{scenarios = list("Scenario 1" = list(MCAR = list(columns = c(1,2,4,6),
                                                 probability = c(0.4,0.4,0.5,0.6)),
                                     MAR = list(columns = c(5,7),
                                                reg.columns = list(c(1,2,3),
                                                                   c(3,6)),
                                                probability.upper = c(0.9,0.8),
                                                probability.midpoint = list(c(0.8,0.25),
                                                                            c(0.75,0.3)),
                                                probability.lower = c(0.1,0.2)),
                                     MNAR = list(columns = c(8,9),
                                                 probability.upper = c(0.9,0.95),
                                                 probability.midpoint = list(c(0.8,0.25),
                                                                             c(0.75,0.3)),
                                                 probability.lower = c(0.1,0.1))),
                 "Scenario 2" = list(MCAR = list(columns = c(1,2,3),
                                                 probability = c(0.5,0.7,0.1))),
                 "Scenario 3" = ...  )
}

In the case that \code{scenarios} is \code{NULL}, one can specify alternatively \code{missingness_parameters} (an argument otherwise ignored).
Via \code{missingness_parameters} one passes parameters for drawing missingness scenarios to \code{data_ImputeBench()}, for details see RMTBM22.
The syntax of \code{missingness_parameters} works very similar to \code{scenarios} as it expects a named list of named list where each
list entry is comprised of at least one and at most three of the lists named \code{"MCAR"}, \code{"MAR"} and/or \code{"MNAR"}, which itself
contain the parameters for the respective missingness mechanisms. Again for more details we refer to RMTBM22 and the
accompanying vignette.

Example for \code{missingness_parameters}:\preformatted{missingness_parameters = list("Parameters 1" = list("MCAR" = list(columns = 1:10,
                                                                  size = 10,
                                                                  probability = 0.25)),
                              "Parameters 2" = list("MCAR" = list(columns = 1:10,
                                                                  size = 10,
                                                                  probability = 0.05),
                                                     "MAR" = list(columns = 2:4,
                                                                  size = 3,
                                                                  size.regressors = c(1,2),
                                                                  probability.upper = 0.9,
                                                                  probability.lower = 0.1,
                                                                  probability.midpoint = c(0.5,0.2)),
                                                      "MNAR" = list(columns = c(5,10),
                                                                    size = 2,
                                                                    probability.upper = 0.05,
                                                                    probability.lower = 0.95,
                                                                    probability.midpoint = c(0.5,0.3))),
                              "Parameters 3" = list("MAR" = list(columns = 6:10,
                                                                 size = 5,
                                                                 size.regressors = c(1,2),
                                                                 probability.upper = 0.9,
                                                                 probability.lower = 0.1,
                                                                 probability.midpoint = c(0.5,0.2)),
                                                    "MNAR" = list(columns = 1:5,
                                                                  size = 5,
                                                                  probability.upper = 0.05,
                                                                  probability.lower = 0.95,
                                                                  probability.midpoint = c(0.5,0.3))),
                              "Parameters 4" = ...                                                      )
}

Note: If \code{scenarios} and \code{missingness_parameters} are \code{NULL} a default scenario is used as specified in the vignette.
}

\subsection{Customization of the Benchmarking Protocol}{

At the moment the argument \code{error_measure} can not be changed as only its default \code{"l2"} is supported.

The argument \code{scaling_robust} expects a positive numeric. It prevents division by 0 when scaling with the error by the standard
deviation. Default of \code{scaling_robust} is \code{0.1}.

To have the error computed for groups of variables one can specify groups via the argument \code{error_grouping} which should be a
list of column indices assigning all columns a group.

Example of \code{error_grouping} (for a data matrix with 10 columns):\preformatted{grouping = list("Group 1" = c(1,2,3,4),
               "Group 2" = c(5,8,9),
               "Group 3" = c(6,7,10))
}

The argument \code{repetitions}, by default set to \code{5}, controls how imputation performance on a scenario is computed. It expects a
positive integer as an input.
}

\subsection{Technical Arguments}{

If \code{TRUE} the argument \code{verbose} lets the consolue put out a statement after each run of each scenario. Default is \code{FALSE}.

Via \code{seed} one can set a seed for the benchmarking. Consider that this seed is not passed to imputation methods and thus for
methods using randomness such as missForest the benchmarking performance does not need to be the same even if the seeds are
equal.

The parameter \code{max_time}, set by default to \code{2592000} (or 30 days), controls the maximal time which imputation methods are
allowed to compute their results.

The parameter \code{output_data} controls whether additional data (to the benchmarking results) should be stored and returned. Default
is \code{TRUE}. Details on the outputs can be found in the vignette and further below.
}
}
\references{
R. Richter, A. Miloschewski, J.F. Tavares, M.M.B. Breteler, S. Mukherjee (2012), Benchmarking Single
Imputation Methods. in preparation.
}
\seealso{
\code{simulation_ImputeBench}, \code{plot_ImputeBench}, \code{missingness_scenario_from_parameters}, \code{simulate_mask}
}
