% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_count_binary.R
\name{determine_count_binary}
\alias{determine_count_binary}
\title{Detect Integer and Binary Data Columns}
\usage{
determine_count_binary(data)
}
\arguments{
\item{data}{A matrix with numerical entries.}
}
\value{
A list of two vectors, the first of which contains the column indices of all integer columns and the second of which contains
the column indices of all binary columns.
}
\description{
Find all columns that consist of only integer values and of those all columns that consist of only entries in 0 and 1.
}
