% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_by_baseline.R
\name{divide_by_baseline}
\alias{divide_by_baseline}
\title{Dividing all Imputation Errors by the Imputation Error of the Baseline}
\usage{
divide_by_baseline(Evaluation)
}
\arguments{
\item{Evaluation}{The output table of \code{data_ImputeBench} or \code{simulation_ImputeBench}.}
}
\value{
The same table with scaled RMSE results
}
\description{
ImputeBench internal function using the $evaluation list entry of the output of \code{data_ImputeBench} or
\code{simulation_ImputeBench} and scaling the imputation errors with the imputation error of the baseline imputation method.
}
