% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_imputation_sim.R
\name{evaluate_imputation_sim}
\alias{evaluate_imputation_sim}
\title{Evaluating an Imputation on Simulated Data}
\usage{
evaluate_imputation_sim(
  data,
  mask,
  imputed_data,
  data.grouping,
  error_measure = "l2",
  scaling_robust = 0.01
)
}
\arguments{
\item{data}{A numeric matrix (ground truth).}

\item{mask}{A matrix with binary entries (0 = missing, 1 = not missing), the (additional) missingness pattern on which the
imputed matrix is compared against the ground truth matrix.}

\item{imputed_data}{An imputed matrix of \code{data}.}

\item{data.grouping}{A vector indicating how long the simulated data groups are. Expected is a vector with six entries, the first
indicates the amount of Gaussian columns, the second the amount of Poisson columns, the third the amount of binary columns,
the fourth the amount of t-distributed columns, the fifth the amount of sine-transformed columns and the sixth the amount
of spline-transformed columns.}

\item{error_measure}{Choose one of the error measures from \code{c("l2")} (At the moment only \code{"l2"} is supported).
Default is \code{"l2"}.}

\item{scaling_robust}{A positive integer, the robustness parameter of the scaling by the standard deviation (to avoid division
by 0). Default is \code{0.01}.}
}
\value{
A list of imputation errors (overall, continuous columns, all six different column types and the log-odds of the binary
columns).
}
\description{
ImputeBench internal function computing the various errors/losses for an imputed matrix. Written for benchmarking
imputation methods on simulated data, i.e. distinguishing errors on Gaussian, Poisson, binary, t-distributed, sine-transformed
and spline-transformed columns.
}
