% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_RMSEscaled_onmask.R
\name{group_RMSEscaled_onmask}
\alias{group_RMSEscaled_onmask}
\title{Grouped and Scaled RMSE for Imputation Error/Loss}
\usage{
group_RMSEscaled_onmask(
  data,
  mask,
  imputed_data,
  error_grouping,
  scaling_type = "StandardDev",
  scaling_robust = 0.01
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries (ground truth).}

\item{mask}{A matrix with binary entries (0 = missing, 1 = not missing), the (additional) missingness pattern on which the
imputed matrix is compared against the ground truth matrix.}

\item{imputed_data}{An imputed numeric matrix of \code{data}.}

\item{error_grouping}{A list of column indices establishing the error groups that should be computed separately.}

\item{scaling_type}{Character string setting the scaling of the data columns, possible values are \code{"StandardDev"} and \code{"IQR"}.
Default is \code{"StandardDev"}.}

\item{scaling_robust}{Positive numeric, robustness parameter avoiding division by close to zero when the columns are scaled by the standard deviation
when computing the imputation error. Default is \code{0.01}.}
}
\value{
A list with the first entry being the overall error/loss and the second entry being the vector of all group-wise errors/losses.
}
\description{
Computing the error/loss of an imputed matrix given its ground truth via a scaled RMSE averaged over multiple column
groups.
}
\details{
For each list entry of \code{error_grouping} the root mean squared error is compute over all column indices of the respective
group where each column is scaled by its respective standard deviation (\verb{scaling_type = "StandardDev}) or its respective
IQR (\code{scaling_type = "IQR"}) - division by zero is avoided via \code{scaling_robust}. The overall error/loss is computed by averaging the
errors/losses over the groups.
}
