% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_training.R
\name{imputation_training}
\alias{imputation_training}
\title{Training the Arguments of Imputation Methods}
\usage{
imputation_training(masked.data, methods, training)
}
\arguments{
\item{masked.data}{A numeric matrix with missing entries given by \code{NA}.}

\item{methods}{A list of functions, each entry taking as input a numeric matrix with missing entries via \code{data} and a list of
parameters via \code{args} and outputting an imputed matrix.}

\item{training}{A list of functions, each entry taking as input a numeric matrix with missing entries via \code{data} and outputting
a list of parameters \code{args}, which is suited to be passed to the corresponding (via indexing) entry in \code{methods}.}
}
\value{
A list of parameters (giving as a list) for each entry in \code{method}.
}
\description{
Internal function of ImputeBench calling the training function for all imputation methods and outputting the list
of trained parameters.
}
