% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_wrapper.R
\name{imputation_wrapper}
\alias{imputation_wrapper}
\title{Imputing the Data by Provided Imputation Methods}
\usage{
imputation_wrapper(
  data,
  methods,
  method.names,
  method_arguments,
  only.count.columns = NULL,
  binary.columns = NULL,
  max.time = 2592000
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{methods}{A list of functions, each entry taking as input a numeric matrix with missing entries via \code{data} and a list of
parameters via \code{args} and outputting an imputed matrix.}

\item{method.names}{A vector of character strings providing names for each entry in \code{methods}.}

\item{method_arguments}{A list of parameter lists to be passed index-wise to the \code{args} argument of the corresponding entry in
\code{methods}.}

\item{only.count.columns}{A vector of indices that correspond to all integer but not binary columns in \code{data}.}

\item{binary.columns}{A vector of indices that correspond to all binary columns in \code{data}.}

\item{max.time}{The maximal amount of seconds an imputation method is allowed to take. Default is \code{2592000}, i.e. 30 days.}
}
\value{
A list of imputed matrices.
}
\description{
Internal ImputeBench function that calling the imputation functions and outputting a list of imputed matrices
}
\details{
Imputation methods are employed over a for-loop and after imputation is performed integer columns are rounded and binary
columns are rounded and set to 0 if below 0 or 1 if above 1.
}
