% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_scenario_from_parameters.R
\name{missingness_scenario_from_parameters}
\alias{missingness_scenario_from_parameters}
\title{Drawing a Missingness Scenario from Missingness Parameters}
\usage{
missingness_scenario_from_parameters(
  nbr_columns,
  missingness_parameters = NULL,
  seed = NULL
)
}
\arguments{
\item{nbr_columns}{The amount of columns of the to be masked numeric matrix (needed for the defaul values).}

\item{missingness_parameters}{A lists of parameters defining the way the missingness scenario is drawn, for details see
below and the vignette. Default is \code{NULL}, for the default choice see the vignette.}

\item{seed}{An integer, setting a seed.}
}
\value{
A missingness scenario.
}
\description{
Given missingness parameters this function draws a missingness scenario.
}
\details{
For the details on the definition of missingness scenarios and the protocol by which they are drawn from missingness
parameter we refer to the accompanying paper RMTBM22. Find below an example of a \code{missingness_parameters} list of parameters that
can be passed to \code{missingness_scenario_from_parameters}:\preformatted{missingness_parameters = list("MCAR" = list(columns = 1:10,
                                            size = 10,
                                            probability = 0.05),
                              "MAR" = list(columns = 2:4,
                                           size = 3,
                                           size.regressors = c(1,2),
                                           probability.upper = 0.9,
                                           probability.lower = 0.1,
                                           probability.midpoint = c(0.5,0.2)),
                              "MNAR" = list(columns = c(5,10),
                                            size = 2,
                                            probability.upper = 0.05,
                                            probability.lower = 0.95,
                                            probability.midpoint = c(0.5,0.3))
}
}
\references{
Robin Richter, Anne Miloschewski, Juliana F. Tavares, Monique M.B. Breteler and Sach Mukherjee, "Benchmarking Single
Imputation Methods", in preparation.
}
\seealso{
\code{data_ImputeBench}
}
