% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1D_data.R
\name{plot_1D_data}
\alias{plot_1D_data}
\title{Plotting Data Distribution (1D)}
\usage{
plot_1D_data(data, clm = 1, title = NULL, binwidth = NULL)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{clm}{An integer, indicating the column to be plotted. Default is \code{1}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}

\item{binwidth}{(1D) A positive numeric controlling the binwidth. Default is \code{NULL}.}
}
\value{
A ggplot2 plot.
}
\description{
Plotting the distribution of given data in 1 dimensions using ggplot2.
}
\details{
Plotting the data distribution of selected columns in \code{clm} of \code{data} as a histogram with customizable \code{binwidth}.
}
