% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1D_imputed.R
\name{plot_1D_imputed}
\alias{plot_1D_imputed}
\title{Plotting Observed and Imputed Data Distributions (1D)}
\usage{
plot_1D_imputed(data, imputed_data, clm = 1, binwidth = NULL, title = NULL)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries (ground truth matrix).}

\item{imputed_data}{An imputed matrix of \code{data}.}

\item{clm}{(1D) An integer, indicating the column to be plotted. Default is \code{1}.}

\item{binwidth}{(1D) A positive numeric controlling the binwidth. Default is \code{NULL}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 plot.
}
\description{
Plotting the distribution of observed and imputed data entries for a given data matrix with missing entries
and an imputation thereof, in 1 dimension using ggplot2.
}
\details{
Plotting the imputed values in \code{clm} of \code{data} against the observed ones in the same column.  In the case that \code{clm}
is continuous, the output is a histogram plot with the imputed values displayed as transparent red bars and the observed
values in filled grey bars. In the case that \code{clm} is discrete, the output is a bar plot color coded for imputed
and observed entries. In the continuous case via \code{binwidth} the width of the bins can be changed.
}
