% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1D_imputed_sim.R
\name{plot_1D_imputed_sim}
\alias{plot_1D_imputed_sim}
\title{Plotting Observed, Missing and Imputed Data Distributions (1D)}
\usage{
plot_1D_imputed_sim(
  data,
  mask,
  imputed_data,
  additional.imputations = NULL,
  imp.names = NULL,
  clm = 1,
  alpha = 1,
  jitter = 0.4,
  title = NULL
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries (ground truth matrix).}

\item{mask}{A matrix with binary entries (0 = missing, 1 = not missing), the (additional) missingness pattern on which the
imputed matrix is compared against the ground truth matrix.}

\item{imputed_data}{An imputed matrix of \code{data}.}

\item{additional.imputations}{A list of imputed matrices to compare different imputation methods. Default is \code{NULL}.}

\item{imp.names}{A vector of character strings providing names to the list of imputation methods in the case
\code{additional.imputations} are passed. The first entry corresponds to \code{imputed_data} while the next entries correspond in their order
to the list entries of \code{additional.imputations}. Default is \code{NULL}.}

\item{clm}{An integer, indicating the column to be plotted. Default is \code{1}.}

\item{alpha}{A positive numeric controlling the transparency of the observed data points. Default is \code{0.25}.}

\item{jitter}{A positive numeric controlling the jitter. Default is \code{0.4}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 plot.
}
\description{
Plotting the distribution of observed, missing and imputed data points for a given data matrix, its missingness
pattern and an imputation thereof, in 1 dimensions using ggplot2.
}
\details{
Plotting the imputed vs. the missing values in \code{clm} of \code{data} including where of interest the observed entries.
In the case that \code{clm} is continuous, the output is a scatter plot of imputed vs. missing values, the line \code{x=y} is drawn
in black. In the case that \code{clm} is discrete, the output is a tile plot of the imputed vs. the missing values, the diagonal
is marked by a red frame around the tiles. In the continuous case via \code{alpha} the alpha can be changed.
}
