% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1D_simulated_missingness.R
\name{plot_1D_simulated_missingness}
\alias{plot_1D_simulated_missingness}
\title{Plotting Observed and Missing Data Distributions (1D)}
\usage{
plot_1D_simulated_missingness(data, mask, clm = 1, title = NULL)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries (ground truth matrix).}

\item{mask}{A matrix with binary entries (0 = missing, 1 = not missing), the (additional) missingness pattern on which the
imputed matrix is compared against the ground truth matrix.}

\item{clm}{An integer, indicating the column to be plotted. Default is \code{1}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 plot.
}
\description{
Plotting the distribution of the observed and missing data entries for a given data matrix and
missingness pattern, in 1 dimension using ggplot2.
}
\details{
Plotting the missingness pattern of \code{clm} of \code{data} against its data distribution. All rows with a (real) missing
value in \code{clm} are left out. In the case that \code{clm} is continuous, the output plot is a box-plot comparing the subdistributions
of missing entries and existing entries. In the case that \code{clm} is discrete, the output plot is a filled bar plot with color
coded missingness (Additionally one finds the total number of missing and existing entries for the given bin on the top and
bottom of the bar).
}
