% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D_data.R
\name{plot_2D_data}
\alias{plot_2D_data}
\title{Plotting Data Distribution (2D)}
\usage{
plot_2D_data(
  data,
  clms = c(1, 2),
  alpha = 0.25,
  jitter = 0.4,
  type = "scatter",
  title = NULL
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{clms}{An integer vector, indicating the columns to be plotted. Default is \code{c(1,2)}.}

\item{alpha}{A positive numeric controlling the transparency of the points. Default is \code{0.25}.}

\item{jitter}{A positive numeric controlling the jitter on the discrete values. Default is \code{0.4}.}

\item{type}{A character string. In the case two discrete variables are selected via \code{clms}, choose between \code{"scatter"} and
\code{"tile"}. Default is \code{"scatter"}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 plot.
}
\description{
Plotting the distribution of given data in 2 dimensions using ggplot2.
}
\details{
Plotting the data distribution of selected columns in \code{clms} of \code{data}.
\subsection{\code{type = "scatter"}}{

The result is a scatter plot with jitter added to discrete variables (customizable via \code{jitter}) with customizable transparency
of the points via \code{alpha}.
}

\subsection{\code{type = "tile"}}{

Only possible if both variables are discrete: The result is a tile plot with labeled tiles.
}
}
