% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D_imputed_sim.R
\name{plot_2D_imputed_sim}
\alias{plot_2D_imputed_sim}
\title{Plotting Observed, Missing and Imputed Data Distributions (2D)}
\usage{
plot_2D_imputed_sim(
  data,
  mask,
  imputed_data,
  alpha = 0.25,
  jitter = 0.4,
  alpha_line = 1,
  clms = c(1, 2),
  type = "scatter",
  title = NULL
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries (ground truth matrix).}

\item{mask}{A matrix with binary entries (0 = missing, 1 = not missing), the (additional) missingness pattern on which the
imputed matrix is compared against the ground truth matrix.}

\item{imputed_data}{An imputed matrix of \code{data}.}

\item{alpha}{A positive numeric controlling the transparency of the observed data points. Default is \code{0.25}.}

\item{jitter}{A positive numeric controlling the jitter on the discrete values. Default is \code{0.4}.}

\item{alpha_line}{A postive numeric controlling the transparency of the lines connecting missing and imputed values.
Default is \code{1}.}

\item{clms}{An integer vector, indicating the columns to be plotted. Default is \code{c(1,2)}.}

\item{type}{A character string. In the case two discrete variables are selected via \code{clms}, choose between \code{"scatter"} and
\code{"tile"}. Default is \code{"scatter"}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 plot.
}
\description{
Plotting the distribution of observed, missing and imputed data points for a given data matrix, its missingness
pattern and an imputation thereof, in 2 dimensions using ggplot2.
}
\details{
Plotting the imputed vs. the missing values in \code{clms} of \code{data} including where of interest the observed entries.
\subsection{\code{type = "scatter"}}{

The output plot is a 2D scattered point plot (with additional jitter in the case that one or both variables of \code{clms} are
discrete) where observed points are blue dots and missing points are red dots, while imputed points are shown as red crosses.
Each missing point is connected to its imputed value (consider that no distinction is made if only the x-value, only the y-value or
both values are missing). The \code{alpha} parameter controls the alpha of the complete points. The \code{jitter} parameter controls
the extra jitter.
}

\subsection{\code{type = "tile"}}{

The output is a tile plot in which for two discrete variables the distribution over the discrete bins is shown for the latent
data (in black at the bottom of each tile) and of the imputed data (in red on the top of each tile). Color coded is the relative
change in value.
}
}
