% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D_real_missingness.R
\name{plot_2D_real_missingness}
\alias{plot_2D_real_missingness}
\title{Plotting the Correlation between Observed Data and the Missingness Pattern (2D)}
\usage{
plot_2D_real_missingness(
  data,
  miss.clm = 1,
  data.clms = c(2, 3),
  alpha = 0.25,
  jitter = 0.4,
  type = "scatter",
  title = NULL
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries (ground truth matrix).}

\item{miss.clm}{An integer, indicating the column whose missingness pattern is to be plotted. Default is \code{1}.}

\item{data.clms}{An integer vector, indicating the columns whose data distribtuion is to be plotted.
Default is \code{c(2,3)}.}

\item{alpha}{A positive numeric controlling the transparency of the observed data points. Default is \code{0.25}.}

\item{jitter}{A positive numeric controlling the jitter on the discrete values. Default is \code{0.4}.}

\item{type}{A character string. In the case one variable is discrete and one is continuous, choose between \code{"scatter"}
and \code{"box"}. Default is \code{"scatter"}. In the case both variables are discrete, choose between \code{"scatter"} and \code{"facets"}.
Default is \code{"scatter"}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 plot.
}
\description{
Plotting the distribution of observed entries of a given data matrix with respect to the missingness pattern of an
additional column of the same matrix, in 2 dimensions using ggplot2.
}
\details{
Plotting the missingness pattern of \code{miss.clm} with respect to the data distribution of \code{data.clms}.
All rows with a missing value in \code{data.clms} are left out.
\subsection{\code{type = "scatter"}}{

The output plot is a 2D scattered point plot (with additional jitter in the case that one or both variables of \code{data.clms} are
discrete) with color coded missingness in \code{miss.clm}. The \code{alpha} parameter controls the alpha of the complete points.
The \code{jitter} parameter controls the extra jitter.
}

\subsection{\code{type = "box"}}{

The output plot is a box plot stratified by the discrete variable of \code{data.clms} on the x-axis and color coded by the missingness
pattern of \code{miss.clm}. Only available if \code{data.clms} consists of a discrete and a continuous variable.
}

\subsection{\code{type = "facets"}}{

The output plot is a filled bar plot faceted by the second variable of \code{data.clms} and color coded by the missingness
pattern of \code{miss.clm}. Only available if \code{data.clms} consists of two discrete variables.
}
}
