% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D_simulated_missingness.R
\name{plot_2D_simulated_missingness}
\alias{plot_2D_simulated_missingness}
\title{Plotting Observed and Missing Data Distributions (2D)}
\usage{
plot_2D_simulated_missingness(
  data,
  mask,
  clms = c(1, 2),
  type = "scatter",
  only.first.entry = FALSE,
  alpha = 0.25,
  jitter = 0.4,
  title = NULL
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries (ground truth matrix).}

\item{mask}{A matrix with binary entries (0 = missing, 1 = not missing), the (additional) missingness pattern on which the
imputed matrix is compared against the ground truth matrix.}

\item{clms}{An integer vector, indicating the columns to be plotted. Default is \code{c(1,2)}.}

\item{type}{A character string. In the case one variable is discrete and the other continuous, choose between \code{"scatter"}
and \code{"box"}. Default is \code{"scatter"}. In the case both variables are discrete, choose between \code{"scatter"} and \code{"facets"}.
Default is \code{"scatter"}.}

\item{only.first.entry}{Boolean, if \code{TRUE} only the missingness pattern of the first entry of \code{clms} is distinguished.
If \code{FALSE} distinction between all possible combination of missing entries between the columns indicated in \code{clms} is made.
Default is \code{FALSE}.}

\item{alpha}{A positive numeric controlling the transparency of the observed data points. Default is \code{0.25}.}

\item{jitter}{A positive numeric controlling the jitter on the discrete values. Default is \code{0.4}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 plot.
}
\description{
Plotting the distribution of the observed and missing data entries for a given data matrix and
missingness pattern, in 2 dimensions using ggplot2.
}
\details{
Plotting the missingness pattern of \code{clms} of \code{data} (depending on \code{only.first.entry} considering only the first entry
of \code{clms} or all) against their data distribution. All rows with a (real) missing value in \code{clms} are left out.
\subsection{\code{type = "scatter"}}{

The output plot is a 2D scattered point plot (with additional jitter in the case that one or both variables of \code{clms} are
discrete) with color coded missingness of the first (if \code{only.first.entry}is \code{TRUE}) or all (if \code{only.first.entry} is \code{FALSE})
columns of \code{clms}. The \code{alpha} parameter controls the alpha of the complete points. The \code{jitter} parameter controls the extra jitter.
}

\subsection{\code{type = "box"}}{

The output plot is a box plot stratified by the discrete variable of \code{clms} on the x-axis and color coded by the missingness
pattern of the first (if \code{only.first.entry}is \code{TRUE}) or all (if \code{only.first.entry} is \code{FALSE})
columns of \code{clms}. Only available if \code{clms} consists of a discrete and a continuous variable.
}

\subsection{\code{type = "facets"}}{

The output plot is a filled bar plot faceted by the second variable of \code{clms} and color coded by the missingness
pattern of the first (if \code{only.first.entry}is \code{TRUE}) or all (if \code{only.first.entry} is \code{FALSE})
columns of \code{clms}. Only available if \code{clms} consists of two discrete variables.
}
}
