% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3D_data.R
\name{plot_3D_data}
\alias{plot_3D_data}
\title{Plotting Data Distribution (3D)}
\usage{
plot_3D_data(
  data,
  clms = c(1, 2, 3),
  alpha = 0.25,
  jitter = 0.4,
  type = "scatter",
  title = NULL
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{clms}{An integer vector, indicating the columns to be plotted. Default is \code{c(1,2,3)}.}

\item{alpha}{A positive numeric controlling the transparency of the points. Default is \code{0.25}.}

\item{jitter}{A positive numeric controlling the jitter on the discrete values. Default is \code{0.4}.}

\item{type}{A character string. In the case the third variable is discrete, choose between \code{"scatter"} and
\code{"facets"}. Default is \code{"scatter"}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 plot (\code{type = facets}) or a plotly plot (\code{type = "scatter"}).
}
\description{
Plotting the distribution of given data in 3 dimensions using ggplot2 and plotly.
}
\details{
Plotting the data distribution of selected columns in \code{clms} of \code{data}.
\subsection{\code{type = "scatter"}}{

The result is a scatter plot with jitter added to discrete variables (customizable via \code{jitter}) with customizable transparency
of the points via \code{alpha}.
}

\subsection{\code{type = "facets"}}{

Only possible if the third column of \code{clms} is discrete. The result is a scatter plot over the first two entries of \code{clms} faceted
over the third entry of \code{clms}. As in the 2D plot jitter is added to discrete variables of the scatter plots and can be controlled
via \code{jitter} and the transparency of the points can be controlled via \code{alpha}.
}
}
