% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3D_imputed.R
\name{plot_3D_imputed}
\alias{plot_3D_imputed}
\title{Plotting Observed and Imputed Data Distributions (3D)}
\usage{
plot_3D_imputed(
  data,
  imputed_data,
  alpha = 0.25,
  jitter = 0.4,
  clms = c(1, 2, 3),
  title = NULL
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries (ground truth matrix).}

\item{imputed_data}{An imputed matrix of \code{data}.}

\item{alpha}{A positive numeric controlling the transparency of the observed data points. Default is \code{0.25}.}

\item{jitter}{A positive numeric controlling the jitter on the discrete values. Default is \code{0.4}.}

\item{clms}{An integer vector, indicating the columns to be plotted. Default is \code{c(1,2,3)}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A plotly plot.
}
\description{
Plotting the distribution of observed and imputed data entries for a given data matrix with missing entries
and an imputation thereof, in 3 dimensions using plotly.
}
\details{
Plotting the imputed values in \code{clms} of \code{data} against the observed ones in the same columns.
The output plot is a 3D scattered point plot (with additional jitter in the case that one or both variables of \code{clms} are
discrete) with color coded which entries have been imputed. The \code{alpha} parameter controls the alpha of the complete points.
The \code{jitter} parameter controls the extra jitter.
}
