% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3D_real_missingness.R
\name{plot_3D_real_missingness}
\alias{plot_3D_real_missingness}
\title{Plotting the Correlation between Observed Data and the Missingness Pattern (3D)}
\usage{
plot_3D_real_missingness(
  data,
  miss.clm = 1,
  data.clms = c(2, 3, 4),
  type = "scatter",
  alpha = 0.25,
  jitter = 0.4,
  title = NULL
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries (ground truth matrix).}

\item{miss.clm}{An integer, indicating the column whose missingness pattern is to be plotted. Default is \code{1}.}

\item{data.clms}{An integer vector, indicating the columns whose data distribtuion is to be plotted.
Default is \code{c(2,3,4)}.}

\item{type}{A character string. In the case the third variable is discrete, choose between \code{"scatter"} and \code{"facets"}.
Default is \code{"scatter"}.}

\item{alpha}{A positive numeric controlling the transparency of the observed data points. Default is \code{0.25}.}

\item{jitter}{A positive numeric controlling the jitter on the discrete values. Default is \code{0.4}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 plot (\code{type = facets}) or a plotly plot (\code{type = "scatter"}).
}
\description{
Plotting the distribution of observed entries of a given data matrix with respect to the missingness pattern of an
additional column of the same matrix, in 3 dimensions using ggplot2 and plotly.
}
\details{
Plotting the missingness pattern of \code{miss.clm} with respect to the data distribution of \code{data.clms}.
All rows with a missing value in \code{data.clms} are left out.
\subsection{\code{type = "scatter"}}{

The output plot is a 3D scatter point plot (with additional jitter in the case that one, two or all variables of \code{data.clms} are
discrete) with color coded missingness in \code{miss.clm}. The \code{alpha} parameter controls the alpha of the complete points.
The \code{jitter} parameter controls the extra jitter.
}

\subsection{\code{type = "facets"}}{

The output plot is a 2D scatter point plot faceted by the third variable of \code{data.clms} and color coded by the missingness
pattern of \code{miss.clm}. Only available if the third column of \code{data.clms} is a discrete variable. The \code{alpha} parameter controls the
alpha of the complete points. The \code{jitter} parameter controls the extra jitter.
}
}
