% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3D_simulated_missingness.R
\name{plot_3D_simulated_missingness}
\alias{plot_3D_simulated_missingness}
\title{Plotting Observed and Missing Data Distributions (3D)}
\usage{
plot_3D_simulated_missingness(
  data,
  mask,
  clms = c(1, 2, 3),
  type = "scatter",
  jitter = 0.4,
  alpha = 0.25,
  title = NULL,
  only.first.entry = FALSE
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries (ground truth matrix).}

\item{mask}{A matrix with binary entries (0 = missing, 1 = not missing), the (additional) missingness pattern on which the
imputed matrix is compared against the ground truth matrix.}

\item{clms}{An integer vector, indicating the columns to be plotted. Default is \code{c(1,2,3)}.}

\item{type}{A character string. In the case the third variable is discrete, choose between \code{"scatter"} and \code{"facets"}.
Default is \code{"scatter"}.}

\item{jitter}{A positive numeric controlling the jitter on the discrete values. Default is \code{0.4}.}

\item{alpha}{A positive numeric controlling the transparency of the observed data points. Default is \code{0.25}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}

\item{only.first.entry}{Boolean, if \code{TRUE} only the missingness pattern of the first entry of \code{clms} is distinguished.
If \code{FALSE} distinction between all possible combination of missing entries between the columns indicated in \code{clms} is made.
Default is \code{FALSE}.}
}
\value{
A ggplot2 plot (\code{type = facets}) or a plotly plot (\code{type = "scatter"}).
}
\description{
Plotting the distribution of the observed and missing data entries for a given data matrix and
missingness pattern, in 3 dimensions using ggplot2 and plotly.
}
\details{
Plotting the missingness pattern of \code{clms} of \code{data} (depending on \code{only.first.entry} considering only the first entry
of \code{clms} or all) against its/their data distribution. All rows with a (real) missing value in \code{clms} are left out.
\subsection{\code{type = "scatter"}}{

The output plot is a 3D scatter point plot (with additional jitter in the case that one, two or all variables of \code{data.clms} are
discrete) with color coded missingness pattern of the first (if \code{only.first.entry}is \code{TRUE}) or all (if \code{only.first.entry} is \code{FALSE})
columns of \code{clms}. The \code{alpha} parameter controls the alpha of the complete points. The \code{jitter} parameter controls the extra jitter.
}

\subsection{\code{type = "facets"}}{

The output plot is a 2D scatter point plot faceted by the third variable of \code{data.clms} with color coded missingness
pattern of the first (if \code{only.first.entry}is \code{TRUE}) or all (if \code{only.first.entry} is \code{FALSE}) columns of \code{clms}.
Only available if the third column of \code{data.clms} is a discrete variable. The \code{alpha} parameter controls the alpha of the complete points.
The \code{jitter} parameter controls the extra jitter.
}
}
