% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amount_missing_entries.R
\name{plot_amount_missing_entries}
\alias{plot_amount_missing_entries}
\title{Amount of Missing Entries Plot}
\usage{
plot_amount_missing_entries(
  data,
  groups = NULL,
  thres_name = 0.05,
  jitter = 0.4,
  alpha = 1,
  seed = NULL,
  y.axis = "percent",
  type = "box",
  title = NULL
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{groups}{A list of column groups. If provided, the output is given group-wise, rather than variable-wise. Default is \code{NULL},
i.e. variable-wise.}

\item{thres_name}{A positive numeric, threshold (percentage of missing entries) whose name of the variables should be displayed?
Only applicable if \code{groups = NULL}. Default is \code{0.05}.}

\item{jitter}{A positive numeric controlling the jitter in horizontal direction. Only applicable if \code{groups = NULL}. Default
is \code{0.4}.}

\item{alpha}{A numeric between 0 and 1 controlling the transparency of the points. Only applicatlbe if \code{groups = NULL}. Default
is 1.}

\item{seed}{An integer to fix a seed of the jitter. ONly applicable if \code{groups = NULL}. Default is \code{NULL}.}

\item{y.axis}{A character string, either \code{"percent"} or \code{"amount"} controlling the labeling of the y-axis either report
relative or total amount of missing entries. Default is \code{"percent"}.}

\item{type}{A character string, one of the following: \code{"box"}, \code{"total"} and \code{"mean"}, controlling the type of plot.
Only applicable if \code{groups} are provided. Default is \code{"box"}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 plot.
}
\description{
Plot the amount of missing entries as a box- or pointplot over groups of columns
}
\details{
Plotting the relative/total amount of missing entries in each column/variable of the given data set or in each group of columns
using ggplot2.
\subsection{Without Column Grouping (\code{groups = NULL})}{

Plotting a point-plot (with jitter in the horizontal direction). Labels and ticks on the y-axis are determined by \code{y.axis}.
Which variables are labeled is determined by \code{thres_name}. Jitter can be controlled via the parameters \code{jitter} and \code{seed},
transparency can be controlled via \code{alpha}. The parameter \code{type} is discarded.
}

\subsection{With Column Grouping}{

For \code{type="box"} plotting a box-plot graphic of the distributions of missing values (relative or total depending on \code{y.axis}) over the
variables of the groups specified in \code{groups}. For \code{type="mean"} plotting a point-plot of the mean values of missing values (relative
or total depending on \code{y.axis}) over the variables of the groups specified in \code{groups}. For \code{type="total"} plotting a point-plot of the
total amount of missing entries in each group (not compatible with \code{y.axis="percent"}). The parameters \code{jitter}, \code{alpha}, \code{thres_name} and
\code{seed} are discarded.
}
}
