% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ImputeBench.R
\name{plot_ImputeBench}
\alias{plot_ImputeBench}
\title{Plotting the Results of ImputeBench}
\usage{
plot_ImputeBench(
  Evaluation,
  title = "Imputation Performance",
  group = NULL,
  x_axis = NULL,
  y_axis = NULL,
  norm_by_baseline = TRUE,
  color_axis_labels = NULL,
  error_type = "overall"
)
}
\arguments{
\item{Evaluation}{The results of \code{data_ImputeBench} or \code{simulation_ImputeBench} given in the data frame \verb{$evaluation}.}

\item{title}{A title for the plot. Default is \code{"Imputation Performance"}.}

\item{group}{A character string. Can be used in the case that \code{Evaluation} is the result of running \code{data_ImputeBench} with given
error grouping (see \code{data_ImputeBench} for details): Specifying \code{group} one can select the imputation performance over an error
group to be plotted by passing its name. Default is \code{NULL} (plotting the overall error).}

\item{x_axis}{A character string. Can be used in the case that \code{Evaulation} is the result of running \code{simulation_ImputeBench}.
One can choose a data or missingness parameter to be used to label the x-axis. For the choices see the accompanying vignette.
Default is \code{NULL}.}

\item{y_axis}{A character string, title of the y-axis. Default is \code{NULL}.}

\item{norm_by_baseline}{Boolean. If \code{FALSE} the total error is reported. If \code{TRUE} the relative error (divided by the baseline
imputation method error) is reported. Default is \code{TRUE}.}

\item{color_axis_labels}{A vector of character strings, labeling the color axis labels. Default is \code{NULL}.}

\item{error_type}{A character string. Can be used in the case that \code{Evaulation} is the result of running \code{simulation_ImputeBench}.
If \code{error_type = "overall"} the usual error is plotted, if one of the following \code{"gauss"}, \code{"pois"}, \code{"binary.mse"},
\code{"binary.log.odds"}, \code{"t"}, \code{"sine"}, \code{"spline"} or \code{"continuous"} is passed via \code{error_type} the errors on the respective
data types are plotted. Default is \code{"overall"}.}
}
\value{
A ggplot2 plot.
}
\description{
Plotting with ggplot2 the results of \code{data_ImputeBench} and \code{simulation_ImputeBench}.
}
\details{
This function plots a boxplot over the different missingness scenarios or data/missingness parameter choices reporting
on the y-axis the imputation error of all imputation methods compared which are distinguished by color.
}
