% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missingness_over_variables.R
\name{plot_missingness_over_variables}
\alias{plot_missingness_over_variables}
\title{Plot Missingness Distribution}
\usage{
plot_missingness_over_variables(
  data,
  groups = NULL,
  x_axis = "percent",
  binwidth = NULL,
  title = NULL
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{groups}{A list of column groups. If provided, the output plot is faceted by the groups. Default is \code{NULL}.}

\item{x_axis}{A character string, either \code{"percent"} or \code{"amount"} controlling the labeling of the x-axis either report
relative or total amount of missing entries. Default is \code{"percent"}.}

\item{binwidth}{A positive numeric controlling the binwidth of the histogram. Default is \code{1/30}.}

\item{title}{A character string controlling the title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 plot.
}
\description{
Plotting the distribution of missing entries over all variables as a ggplot2 histogram.
}
\details{
Plotting the histogram over the amount of missing entries in the variables. When specifying \code{groups} the histogram
is spread over multiple facets.
}
