% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sigmoid.R
\name{plot_sigmoid}
\alias{plot_sigmoid}
\title{Plot Piecewise-Sigmoid Function}
\usage{
plot_sigmoid(
  probability.upper = 0.8,
  probability.lower = 0.2,
  probability.midpoint = c(0.5, 0.5)
)
}
\arguments{
\item{probability.upper}{A numeric between 0 and 1. Default is \code{0.8}.}

\item{probability.lower}{A numeric between 0 and 1. Default is \code{0.2}.}

\item{probability.midpoint}{A numeric vector with entries between 0 and 1. Default is \code{c(0.5,0.5)}}
}
\value{
Outputs a plot of the piecewise-sigmoid function.
}
\description{
Visualizing the piecewise-sigmoid function used to construct MAR and MNAR missingness patterns in \code{data_ImputeBench}
and \code{simulation_ImputeBench}.
}
\details{
For details on the construction of the piecewise-sigmoid see the accompanying vignette or the paper RMTBM22.
}
\references{
Robin Richter, Anne Miloschewski, Juliana F. Tavares, Monique M.B. Breteler and Sach Mukherjee, "Benchmarking Single
Imputation Methods", in preparation.
}
