% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSEscaled_onmask.R
\name{RMSEscaled_onmask}
\alias{RMSEscaled_onmask}
\title{Scaled RMSE for Imputation Error/Loss}
\usage{
RMSEscaled_onmask(
  data,
  imputed_data,
  scaling_type = "StandardDev",
  mask,
  reduced.output = FALSE,
  scaling_robust = 0.01
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries (ground truth).}

\item{imputed_data}{An imputed numeric matrix of \code{data}.}

\item{scaling_type}{Character string setting the scaling of the data columns, possible values are \code{"StandardDev"} and \code{"IQR"}.
Default is \code{"StandardDev"}.}

\item{mask}{A matrix with binary entries (0 = missing, 1 = not missing), the (additional) missingness pattern on which the
imputed matrix is compared against the ground truth matrix.}

\item{reduced.output}{Boolean, controlling whether the number of missing entries should be outputted as well. Default is \code{FALSE}.}

\item{scaling_robust}{Positive numeric, robustness parameter avoiding division by close to zero when the columns are scaled by the standard deviation
when computing the imputation error. Default is \code{0.01}.}
}
\value{
A list with the first entry being the scaled RMSE and the second entry the number of imputed matrix entries
(\code{reduced.output = FALSE}) or the RMSE (\code{reduced.output = TRUE}).
}
\description{
Computing the error/loss of an imputed matrix given its ground truth via a scaled RMSE.
}
\details{
The root mean squared error is compute over all missing entries according to \code{mask} that are not \code{NA} in \code{data} where each
column is scaled by its respective standard deviation (\verb{scaling_type = "StandardDev}) or its respective IQR (\code{scaling_type = "IQR"})
\itemize{
\item division by zero is avoided via \code{scaling_robust}.
}
}
