% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_simulation_study_oneRun.R
\name{scenario_simulation_study_oneRun}
\alias{scenario_simulation_study_oneRun}
\title{One Run of the Benchmarking of \code{data_ImputeBench}}
\usage{
scenario_simulation_study_oneRun(
  data,
  methods,
  method.names,
  training,
  scenario,
  scenario.name,
  repetitions = 5,
  error_grouping,
  error_measure = "l2",
  scaling_robust = 0.01,
  retrain.always = FALSE,
  verbose = FALSE,
  seed = NULL,
  run.nbr = 1,
  max.time = 2592000,
  output_data = TRUE
)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{methods}{A list of functions, taking as an input a numeric matrix with missing entries via an argument \code{data} and a list of
parameters via an argument \code{args}. Default is \code{NULL}, i.e. no additional methods added to the benchmarking.}

\item{method.names}{A vector of character strings, corresponding to the names of \code{methods}.}

\item{training}{A list of functions - which need to be of the same length as the \code{methods} list -, that take as an input
a data matrix with missing entries via an argument \code{data} and outputs the parameters for the respective method via as a list \code{args}.}

\item{scenario}{A list constituting a missingness scenario (parameters thereof), for details and an example see the vignette or
the documentation of \code{data_ImputeBench}.}

\item{scenario.name}{A character string giving the name of the \code{scenario}.}

\item{repetitions}{Integer determining how often an additional missingness pattern according to one entry of \code{scenarios} or  \code{missingness_parameters}
is drawn. Default is \code{5}.}

\item{error_grouping}{A list of indice vectors defining column groups. Errors for each groups are collected and the overall error
is computed by the mean error over all groups (weighting all groups the same irrespective of the number of columns they contain).}

\item{error_measure}{Character string that must be \code{"l2"} (up to now only error measure supported). Default is \code{"l2"}.}

\item{scaling_robust}{Positive numeric, robustness parameter avoiding division by close to zero when the columns are scaled by the standard deviation
when computing the imputation error. Default is \code{0.01}.}

\item{retrain.always}{Boolean, should parameters for each imputation method be learned for every run or only every scenario/parameter choice?
Default is \code{FALSE}.}

\item{verbose}{Boolean, determines if an output to the console is made each time a run is finished. Default is \code{FALSE}.}

\item{seed}{Integer or \code{NULL}, sets a seed. Default is \code{NULL}.}

\item{run.nbr}{Integer indicating the number of run in the simulation study. Default is \code{1}.}

\item{max.time}{The maximal amount of seconds an imputation method is allowed to take. Default is \code{2592000}, i.e. 30 days.}

\item{output_data}{Boolean, should additional data be collected, i.e. drawn missingness pattern? Default is \code{TRUE}.}
}
\value{
Results of one scenario/missingness parameters of \code{data_ImputeBench}.
}
\description{
Internal function of ImputeBench computing imputation error for a single scenario/missingness parameters within
the benchmarking protocol of \code{data_ImputeBench}.
}
\references{
R. Richter, A. Miloschewski, J.F. Tavares, M.M.B. Breteler, S. Mukherjee (2012), Benchmarking Single
Imputation Methods. in preparation.
}
