% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Data Simulation Protocol of ImputeBench}
\usage{
simulate_data(
  data_parameters = list(data.size = 100, data.type = c(5, 5, 5, 5, 5, 5), data.shape =
    list(gauss.mean = c(0, 1), gauss.sd = c(1, 2), pois.lambda = c(1), binary.limits =
    c(0.2, 0.8), t.mean = c(0, 1), t.limits = c(1, 20), nl.mean = c(0, 1), nl.limits =
    c(10, 30), nl.freq = c(10, 20, 15, 1), sp.points = 5, incl.group = 0), data.groups =
    1, data.groups.distance = c(0.5, 1.5), data.non.linearity = 0, data.spars = 0.4,
    data.low.dim = 5, data.low.dim.groups = 1, data.low.dim.parameters =
    list(trunc.parameter = c(0.5, 10,      2, 0.5), mean = 1, error.std = 0.05)),
  seed = NULL
)
}
\arguments{
\item{data_parameters}{A named list of data parameters, for details see below and the accompanying vignette. Default is given
further below.}

\item{seed}{An integer, sets a seed for the simulation. Default is \code{NULL}.}
}
\value{
A numeric matrix.
}
\description{
Simulating data according to the data simulation protocol of RMTBM22.
}
\details{
The parameter \code{data_parameters} needs to be a list specifying the data parameters of the data simulation protocol
defined in RMTBM22. Not all parameters have to be specified, missing ones are replaced by default ones. For a list and discussion
of the parameters we refer to the accompanying vignette and RMTBM22. Here we just give an example that also constitutes the
default of \code{data_parameters}:\preformatted{data_parameters = list(data.size = 100,
                       data.type = c(5,5,5,5,5,5),
                       data.shape = list(gauss.mean = c(0,1),
                                         gauss.sd = c(1,2),
                                         pois.lambda = c(1),
                                         binary.limits = c(0.2,0.8),
                                         t.mean = c(0,1),
                                         t.limits = c(1,20),
                                         nl.mean = c(0,1),
                                         nl.limits = c(10,30),
                                         nl.freq = c(10,20,15,1),
                                         sp.points = 5,
                                         incl.group = 0),
                       data.groups = 1,
                       data.groups.distance = c(0.5,1.5),
                       data.non.linearity = 0,
                       data.spars = 0.4,
                       data.low.dim = 5,
                       data.low.dim.groups = 1,
                       data.low.dim.parameters = list(trunc.parameter = c(0.5,10,2,0.5),
                                                      mean = 1,
                                                      error.std = 0.05))
}
}
\references{
R. Richter, A. Miloschewski, J.F. Tavares, M.M.B. Breteler, S. Mukherjee (2012), Benchmarking Single
Imputation Methods. in preparation.
}
\seealso{
\code{data_ImputeBench}
}
