% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mask.R
\name{simulate_mask}
\alias{simulate_mask}
\title{Missingness Pattern Simulation Protocol of ImputeBench}
\usage{
simulate_mask(data, scenario = NULL, seed = NULL)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{scenario}{A named list defining the missingness scenario, for details see below and the accompanying vignette.
Default is given further below.}

\item{seed}{An integer, sets a seed for the simulation. Default is \code{NULL}.}
}
\value{
Binary missingness pattern matrix for the given data.
}
\description{
Simulating a missingness pattern given a missingness scenario according to the missingness pattern simulation
protocol of RMTBM22.
}
\details{
The parameter \code{scenario} needs to be a list specifying the missingness scenario (as defined in RMTBM22) guiding
the missingness pattern simulation protocol defined in RMTBM22. Not all parameters have to be specified, missing ones are
replaced by default ones. For a list and discussion of the parameters we refer to the accompanying vignette and RMTBM22.
Here we just give an example of \code{scenario} for a data matrix with at least 9 columns:\preformatted{scenario = list(MCAR = list(columns = c(1,2,4,6),
                            probability = c(0.4,0.4,0.5,0.6)),
                MAR = list(columns = c(5,7),
                           reg.columns = list(c(1,2,3),
                                              c(3,6)),
                           probability.upper = c(0.9,0.8),
                           probability.midpoint = list(c(0.8,0.25),
                                                       c(0.75,0.3)),
                           probability.lower = c(0.1,0.2)),
                MNAR = list(columns = c(8,9),
                            probability.upper = c(0.9,0.95),
                            probability.midpoint = list(c(0.8,0.25),
                                                        c(0.75,0.3)),
                            probability.lower = c(0.1,0.1))),

}

The default of \code{scenario} is given by the missingness pattern present in \code{data}: For each column the percentage of missing
entries is calculated. In the case there are missing entries the columns index is written into \code{MCAR$columns} and the percentage
of missing entries is denoted in the same-index entry of \code{MCAR$probability}.
}
\references{
R. Richter, A. Miloschewski, J.F. Tavares, M.M.B. Breteler, S. Mukherjee (2012), Benchmarking Single
Imputation Methods. in preparation.
}
