% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_KNN.R
\name{train_KNN}
\alias{train_KNN}
\title{Training k-nn Impute via Bisection}
\usage{
train_KNN(data)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}
}
\value{
A list \code{args} consisting of one positive integer entry named \verb{$k}.
}
\description{
Training of the parameter \code{k} of the k-nn imputation exported from the \code{VIM} package by a bisection algorithm
(called within the wrapper function \code{wKNN}).
}
\seealso{
\code{wKNN}, \code{RMSEscaled_onmask}, \code{simulate_mask}
}
