% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_MICE.R
\name{train_MICE}
\alias{train_MICE}
\title{Setting up MICE}
\usage{
train_MICE(data)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}
}
\value{
A list \code{args} consisting of the numeric entry between 0 and 1 called \verb{$mincor}, the positive integer entry called
\verb{$m} and the positive integer entry called \verb{$maxit}.
}
\description{
Internal function of ImputeBench setting the MICE parameters.
}
\seealso{
\code{wMICE}
}
