% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_missF.R
\name{train_missF}
\alias{train_missF}
\title{Setting up missForest}
\usage{
train_missF(data)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}
}
\value{
A list \code{args} consisting of one positive integer entry called \verb{$ntree}.
}
\description{
Internal function of ImputeBench setting the missForest parameters.
}
\seealso{
\code{wmissF}
}
