% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sImpute.R
\name{train_sImpute}
\alias{train_sImpute}
\title{Training Soft Impute via Bisection}
\usage{
train_sImpute(data)
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}
}
\value{
A list \code{args} consisting of the positive numeric entry named \verb{$lambda}, the positive integer entry \verb{$rank.max} and
the character string entry \verb{$type}.
}
\description{
Training of the parameter \code{lambda} of soft impute of the \code{softImpute} package by a bisection algorithm
(called within the wrapper function \code{wsImpute}). Moreover, setting of the remaining parameters.
}
\seealso{
\code{wsImpute}, \code{RMSEscaled_onmask}, \code{simulate_mask}
}
