% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivial_impute.R
\name{trivial.impute}
\alias{trivial.impute}
\title{Baseline Imputation Method of ImputeBench}
\usage{
trivial.impute(data, args = list(method = "median"))
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{args}{A list with one entry \verb{$method}. Default is \code{list(method = "median")}.}
}
\value{
A numeric matrix without missing entries
}
\description{
The baseline imputation method included in ImputeBench is supporting at the moment mean and median imputation.
}
