% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wKNN.R
\name{wKNN}
\alias{wKNN}
\title{Wrapper of k-nn Impute for ImputeBench}
\usage{
wKNN(data, args = list(k = 5))
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{args}{A list with one integer entry \verb{$k}. Default is \code{list(k = 5)}.}
}
\value{
A numeric matrix without missing entries
}
\description{
Wrapper of k-nn impute of the \code{VIM} package to included it as a default imputation method in ImputeBench.
}
\references{
A. Kowarik, M. Templ (2016) Imputation with R package VIM. Journal of Statistical Software, 74(7), 1-16.
}
\seealso{
\code{train_KNN}
}
