% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wMICE.R
\name{wMICE}
\alias{wMICE}
\title{Wrapper of MICE for ImputeBench}
\usage{
wMICE(data, args = list(mincor = NULL, m = 5, maxit = 5))
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{args}{A list with three entries: A numeric between 0 and 1 called \verb{$mincor}, a positive integer called \verb{$m}
and a positive integer called \verb{$maxit}. Default is \code{list(mincor = NULL, m = 5, maxit = 5)}.}
}
\value{
A numeric matrix without missing entries
}
\description{
Wrapper of the MICE imputation function of the \code{mice} package to included it as a default imputation method
in ImputeBench.
}
\references{
S. van Buuren, K. Groothuis-Oudshoorn (2011) \code{mice}: Multivariate Imputation by Chained Equations.
Journal of Statistical Software, 45(3), 1-67.
}
