% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmissF.R
\name{wmissF}
\alias{wmissF}
\title{Wrapper of missForest for ImputeBench}
\usage{
wmissF(data, args = list(ntree = 100))
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{args}{A list with one positive integer entry called \verb{$ntree}. Default is \code{list(ntree = 100)}.}
}
\value{
A numeric matrix without missing entries
}
\description{
Wrapper of the missForest imputation function of the \code{missForest} package to included it as a default imputation
method in ImputeBench.
}
\references{
D.J. Stekhoven, P. Buehlmann (2012) \code{MissForest} - nonparametric missing value imputation for mixed-type data.
Bioinformatics, 28(1), 112-118.
}
