% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsImpute.R
\name{wsImpute}
\alias{wsImpute}
\title{Wrapper of Soft Impute for ImputeBench}
\usage{
wsImpute(data, args = list(rank.max = 50, lambda = 2, type = "svd"))
}
\arguments{
\item{data}{A matrix with numeric entries or \code{NA} entries marking missing entries.}

\item{args}{A list with three entries: A positive integer entry called \verb{$rank.max}, a positive numeric called \verb{$lambda}
and a character string called \verb{$type}. Default is \code{list(rank.max = 50, lambda = 2, type = "svd")}.}
}
\value{
A numeric matrix without missing entries
}
\description{
Wrapper of soft impute of the \code{softImpute} package to included it after data standardization (see details)
as a default imputation method in ImputeBench.
}
\details{
To improve the performance of soft impute, before imputation the data is standardized and after imputation the data is
transformed back by multiplication with the pre-imputation column standard deviation and by adding the pre-imputation column mean.
}
\references{
R. Mazumder, T. Hastie, R. Tibshirani (2010) Spectral Regularization Algorithms for Learning Large Incomplete
Matrices. Journal of Machine Learning Research, 11, 2287-2322.
}
